test bug fixes
